requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
gcsfs
google-cloud-aiplatform
kfp